frontend: Add tests for target="_blank" links and fix one header link
Fix one header link's rel attribute to pass the new tests.

Also move Dropdown to header.jsx and rename to MenuDropdown because it's
implementation is actually specific to the header menus.
