+++ installer/frontend/components/ui.jsx
@@ -580,28 +580,6 @@ export class DropdownMixin extends React.PureComponent {
   }
 }
 
-export class Dropdown extends DropdownMixin {
-  render () {
-    const {active} = this.state;
-    const {items, header} = this.props;
-
-    const children = _.map(items, (href, title) => {
-      return <li className="tectonic-dropdown-menu-item" key={title}>
-        <a className="tectonic-dropdown-menu-item__link" href={href} rel="noopener" target="_blank">{title}</a>
-      </li>;
-    });
-
-    return (
-      <div ref={el => this.dropdownElement = el} className="dropdown" onClick={this.toggle}>
-        <a className="tectonic-dropdown-menu-title">{header}&nbsp;&nbsp;<i className="fa fa-angle-down"></i></a>
-        <ul className="dropdown-menu tectonic-dropdown-menu" style={{display: active ? 'block' : 'none'}}>
-          {children}
-        </ul>
-      </div>
-    );
-  }
-}
-
 export class DropdownInline extends DropdownMixin {
   render () {
     const {active} = this.state;
